apply indenting instructions, correctly detect outdent
Outdents only happen when the node is the indenting node (the last in
the parent-wise path). They do not indent when they're in the middle.

To avoid them from outdenting in the middle, we need to track if the
node we're visiting for an indent instruction is the last node in the path